+++ generators/rest-adapter-core/src/main/java/uk/gov/justice/services/adapter/rest/application/CommonProviders.java
@@ -1,6 +1,7 @@
 package uk.gov.justice.services.adapter.rest.application;
 
 import uk.gov.justice.services.adapter.rest.cors.CorsFeature;
+import uk.gov.justice.services.adapter.rest.filter.LoggerRequestDataFilter;
 import uk.gov.justice.services.adapter.rest.interceptor.JsonSchemaValidationInterceptor;
 import uk.gov.justice.services.adapter.rest.mapper.BadRequestExceptionMapper;
 import uk.gov.justice.services.adapter.rest.mapper.ForbiddenRequestExceptionMapper;
@@ -21,6 +22,7 @@ public class CommonProviders {
         final HashSet<Class<?>> classes = new HashSet<>();
         classes.add(BadRequestExceptionMapper.class);
         classes.add(ForbiddenRequestExceptionMapper.class);
+        classes.add(LoggerRequestDataFilter.class);
         classes.add(JsonSchemaValidationInterceptor.class);
         classes.add(CorsFeature.class);
         return classes;